fix: sync source-to-package lit tests with card-level SPMD contract
Four Tools/Runtime lit failures were stale since 76f68e29 moved the SPMD
layer from 16-rank to card-level semantics:

- Reference capture now uses f16 (the target verifier rejects f32 GEMM),
  aligning the source-to-package vertical with the repository default
  dtype policy; the capture contract mock follows.
- Structured-tensor-program capture switches from the 16-device sharded
  exporter mode to the card-level reference mode.
- The materialize-execution-mesh pass pipeline passes an explicit shape
  (the default tile-count fallback was removed).
- Manifest assertions sync the current interface: input role_index
  numbering, f16 byte widths, and the wafer-run manifest error message.

The two source-to-package compile tests run with --optimization-policy=none
(the Q49.P baseline gate); the default search policy also passes for the
f16 reference program but is out of this task's gate.

Tools/Runtime lit: 33/38 (5 unsupported, 0 failed). Default lit gate
216/216; affected unit suites 70/70; three trees build clean.

+++ test/Tools/Inputs/wafer_pytorch_xla_capture.py
@@ -795,10 +795,10 @@ def _make_reference_matmul_module(torch_module: Any, size: int) -> Any:
         def __init__(self):
             super().__init__()
             self.weight = torch_module.nn.Parameter(
-                torch_module.empty(size, size, dtype=torch_module.float32)
+                torch_module.empty(size, size, dtype=torch_module.float16)
             )
             self.bias = torch_module.nn.Parameter(
-                torch_module.empty(size, dtype=torch_module.float32)
+                torch_module.empty(size, dtype=torch_module.float16)
             )
 
         def forward(self, x):
@@ -1867,7 +1867,7 @@ def emit_reference_stablehlo_program(
     else:
         reference_module = reference_module_factory()
     reference_module.eval()
-    input_tensor = torch_module.empty(size, size, dtype=torch_module.float32)
+    input_tensor = torch_module.empty(size, size, dtype=torch_module.float16)
 
     options = stablehlo_module.StableHLOExportOptions()
     options.export_weights = True
